Move common read&write methods into abstract_comm.py
