+++ instruments/instruments/abstract_instruments/comm/file_communicator.py
@@ -108,18 +108,6 @@ class FileCommunicator(io.IOBase, AbstractCommunicator):
         except IOError as e:
             logger.warn("Failed to close file, exception: %s", repr(e))
 
-    def read(self, size=-1, encoding="utf-8"):
-        """
-        Read bytes in from the file, then decode it to a string using the
-        specified encoding.
-
-        :param int size: The number of bytes to be read in from the file
-        :param str encoding: The encoding with which the read in bytes will
-            be decoded with.
-        :rtype: `str`
-        """
-        return self.read_raw(size).decode(encoding)
-
     def read_raw(self, size=-1):
         """
         Read bytes in from the file.
@@ -141,17 +129,6 @@ class FileCommunicator(io.IOBase, AbstractCommunicator):
             return result
         else:
             raise ValueError("Must read a positive value of characters.")
-
-    def write(self, msg, encoding="utf-8"):
-        """
-        Write a string to the file. The string will be encoded into bytes
-        using the specified encoding scheme.
-
-        :param str msg: String to be written to file
-        :param str encoding: Encoding to apply to ``msg`` to convert it to
-            bytes
-        """
-        self.write_raw(msg.encode(encoding))
 
     def write_raw(self, msg):
         """
