+++ instruments/instruments/abstract_instruments/comm/usb_communicator.py
@@ -92,18 +92,6 @@ class USBCommunicator(io.IOBase, AbstractCommunicator):
     def read(self, size=-1, encoding="utf-8"):
         raise NotImplementedError
 
-    def write(self, msg, encoding="utf-8"):
-        """
-        Write a string to hte raw USB connection object. The string is
-        converted to bytes using the provided encoding method.
-
-        :param str msg: String to be sent to the instrument over the raw USB
-            connection
-        :param str encoding: Encoding to apply on msg to convert the message
-            into bytes
-        """
-        self.write_raw(msg.encode(encoding))
-
     def write_raw(self, msg):
         """
         Write bytes to the raw usb connection object.
